Renamed package director
