+++ unit/core/search/client/test_search_client.py
@@ -2,7 +2,7 @@ import asyncio
 
 from core.search.client import SearchClient
 
-from unit.utils.elasticsearch_test_case import ElasticsearchTestCase
+from unit.mocks.elasticsearch_test_case import ElasticsearchTestCase
 
 
 class SearchClientTestCase(ElasticsearchTestCase):
