+++ unit/ons/search/response/test_ons_response.py
@@ -2,7 +2,7 @@
 Tests the ONSResponse class
 """
 from unit.ons.search.test_utils import SearchTestUtils
-from unit.utils.elasticsearch_test_case import ElasticsearchTestCase
+from unit.mocks.elasticsearch_test_case import ElasticsearchTestCase
 
 from ons.search.response.search_result import SearchResult
 from ons.search.response.ons_response import ONSResponse
